Added frontend components and updated backend for DigiLockerFinal

+++ springapp/src/main/java/com/examly/springapp/repository/DocumentRepository.java
@@ -3,6 +3,6 @@ package com.examly.springapp.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.examly.springapp.model.Document;
 
-public interface DocumentRepository extends JpaRepository<Document,Integer> {
+public interface DocumentRepository extends JpaRepository<Document,Long> {
     
 }
