+++ springapp/src/main/java/com/examly/springapp/repository/FolderRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.examly.springapp.model.Folder;
 
-public interface FolderRepository extends JpaRepository<Folder,Integer>{
+public interface FolderRepository extends JpaRepository<Folder,Long>{
     
 }
